move replay converters to common
