+++ client/src/BattleRoom.lua
@@ -5,7 +5,7 @@ local sceneManager = require("client.src.scenes.sceneManager")
 local GameModes = require("common.engine.GameModes")
 local class = require("common.lib.class")
 local ServerMessages = require("client.src.network.ServerMessages")
-local ReplayV1 = require("client.src.replayV1")
+local ReplayV1 = require("common.engine.replayV1")
 local Signal = require("common.lib.signal")
 local MessageTransition = require("client.src.scenes.Transitions.MessageTransition")
 local ModController = require("client.src.mods.ModController")
